+++ src/main/java/com/example/demofx/controller/PatientController.java
@@ -14,6 +14,7 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
 
 import java.net.URL;
 import java.util.ArrayList;
@@ -172,23 +173,35 @@ public class PatientController implements Initializable {
                 refrechLayout();
                 dialogsController.openInfo("تم عملية الإضافة بنجاح");
             }catch (Exception e){
-                Utils.trackingException(e,"حدث خطأ في عملية الإضافة",dialogsController);
+                new Alert(Alert.AlertType.ERROR, "خطأ في إدخال إتملم العملية  حاول مرة اخرى").show();
             }
         });
         delete.setOnAction(actionEvent -> {
-            currentPage = table.getCurrentPage();
-            context.delete(PATIENT).where(PATIENT.ID.eq(ID)).execute();
-            refrechLayout();
+            if (ID != 0) {
+                try{
+                    currentPage = table.getCurrentPage();
+                    context.delete(PATIENT).where(PATIENT.ID.eq(ID)).execute();
+                    refrechLayout();
+                }catch (Exception e ){
+                    new Alert(Alert.AlertType.ERROR, "خطأ في  إتمام العملية  حاول مرة اخرى").show();
+                }
+            }else{
+                new Alert(Alert.AlertType.ERROR, "خطأ في  إتمام العملية  حاول مرة اخرى").show();
+            }
         });
         update.setOnAction(actionEvent -> {
-            try {
-                PatientRecord patientRecord=initRecord();
-                patientRecord.setId(ID);
-                patientRecord.update();
-                refrechLayout();
-                dialogsController.openInfo("تم عملية التعديل بنجاح");
-            }catch (Exception e){
-                Utils.trackingException(e,"حدث خطأ في عملية الإضافة",dialogsController);
+            if (ID != 0) {
+                try {
+                    PatientRecord patientRecord=initRecord();
+                    patientRecord.setId(ID);
+                    patientRecord.update();
+                    refrechLayout();
+                    dialogsController.openInfo("تم عملية التعديل بنجاح");
+                }catch (Exception e){
+                    new Alert(Alert.AlertType.ERROR, "خطأ في  إتمام العملية  حاول مرة اخرى").show();
+                }
+            }else{
+                new Alert(Alert.AlertType.ERROR, "خطأ في  إتمام العملية  حاول مرة اخرى").show();
             }
         });
     }
